requirements.txt
pandas
numpy
matplotlib
streamlit
seaborn
plotly
pdfminer.six
pdfminer.six[image]